Replace sickle_ui recommendation with bevy_cobweb_ui (#315)
Sickle Ui will no longer be maintained with bevy 0.14 being the last
supported release. With the imminent release of bevy 0.15 and the fact
that there is no support for Sickle Ui it does not make sense to
continue recommending it. See the sickle_ui repo:
https://github.com/UmbraLuminosa/sickle_ui

[Bevy Cobweb Ui ](https://github.com/UkoeHB/bevy_cobweb_ui) is based on
sickle ui, provides a lot more functionality, and the dev has stated
they intend to continue supporting it past sickle ui's death. For the
same reasons that sickle ui was recommended it seems smart to recommend
bevy cobweb ui for the foreseeable future.

---------

Co-authored-by: Ben Frankel <ben.frankel7@gmail.com>