Switch from `trunk` to `bevy run web` (#312)
# Objective

Closes #297.

This PR prepares the project to use the Bevy CLI, mainly the `bevy run`
command.
It simplifies especially running for the web target, which previously
required use of `trunk` (and a bunch of additional configuration).

# Solution

- Remove custom web assets. Use the defaults provided by Bevy CLI
instead.
- Remove custom compilation profiles for the web build. The CLI will
automatically use custom `web` and `web-release` profiles which use
sensible defaults for web builds.
- Adjust the release CI workflow to use the Bevy CLI with `--bundle`
instead of `trunk`.
- Update the README to use Bevy CLI instead of `trunk`.

Note that this introduces a "regression" to the user that they now have
to specify `--no-default-features`, which was automatically passed to
`trunk`.
I created <https://github.com/TheBevyFlock/bevy_cli/issues/250> so we
can do that in Bevy CLI as well.

# Testing

1. Install the Bevy CLI:
    ```
cargo install --git https://github.com/TheBevyFlock/bevy_cli --locked
bevy_cli
    ```
2. Checkout this branch: `gh pr checkout 312` (with the GitHub CLI)
3. Run the app in the browser:
    ```
    bevy run --no-default-features web --open
    ```

---------

Co-authored-by: Ben Frankel <ben.frankel7@gmail.com>