Prefer `with_child` over `with_children` (#344)
`with_child` was added in 0.15.

I also renamed the closure argument in `with_children` from `children`
to `parent`, because that's the most common name used in `bevy` itself
(there are 0 instances of `children`; the runner-up is `builder`).

I know 0.16 adds the new `children![]` macro, which may be more
idiomatic, but I'm not sure and IDK how to use it yet :)