Add in RustRover/Jetbrains Run Configurations (#360)
As seen in #333 there are some potential issues with running/supporting
RustRover and other Jetbrains IDEs in that the default options that are
provided are not necessarily what should actually be used.

In Visual Studio Code, this is currently solved through the
`.vscode/tasks.json` file. The configurations provided here are based
off of what was in that file, and should be mostly equivalent.

This adds in four shell command Build Configurations, using the bevy
shell command. They are put under a Group called "Build", which is done
to approximate how the vscode tasks work.