Add in a Cargo Run Config for local debugging. (#362)
The current Shell Run Configurations currently have an issue where they
don't allow you to debug from RustRover, so this adds in a singular Run
Configuration in Rust Rover that uses cargo that would allow you to
debug.

This is a bit of an addendum to #360 as you can't really debug if just
use those run configurations.

Note that I *can* add in more as well, and at this time, this PR doesn't
allow for web debugging either (although, admittedly, I'm not sure how
web debugging would work).

I've named this `Run Native Debug` so that it's somewhat clear this is
for debugging.

Co-authored-by: Ben Frankel <ben.frankel7@gmail.com>