Implement a settings menu (#374)
![image](https://github.com/user-attachments/assets/263ae947-e290-41c7-b5aa-e7f8c584ccea)

Fixes #213 by adding a menu for volume control.

I know this was deemed out of scope a long time ago, but I want to
resurrect that discussion, mainly for the sake of accessibility. I'd
like to live in a world where Bevy jam games come with accessibility
options (e.g. mouse acceleration, FOV, etc.) out of the box, and IMO
that is only plausible if we give devs and users a settings menu for
that.
The historical answer to this was to add it to the Bevy examples for
users to look up, but
- a) we all know that a jam is on a tight deadline, so I don't expect
users to spend extra time on this.
- b) Bevy should imo be as accessible out of the box as possible, and I
want to extend that to this template.
- c) a settings menu is something that pretty much every single game in
existence will profit from, even if only to allow audio control for the
players. Many browsers don't even allow lowering the volume in a tab
without muting it completely.

The other worry was that this would bloat up the template. To address
this:

- d) the boilerplate for the settings screen is largely contained in the
settings module, so it's supremely easy to rip it out if you don't want
it.
- e) I intentionally wrote the code so that no new abstraction is
introduced. Users don't need to know about this, they are free to
entirely ignore it.

## Future work

- If this is merged, a future PR can make the menu accessible in-game
through pausing. I left that task intentionally to a followup, as its
implementation is largely orthogonal to this.
- We could also add new options to this menu, like screen resolution,
MSAA samples, etc.