Make web builds more robust (#383)
Otherwise, we might get this:


![image](https://github.com/user-attachments/assets/6a4106dd-63bf-4840-a11b-fdc7eeaca8ff)


This error started cropping up in my 0.16 upgrades, dunno why. It
happens on specific uses of `rand`, including through dependencies, so
this might be very subtle for a user, especially since the error message
is less then helpful.

Let's just remove that potential footgun :) I don't know how this
interacts with the `RUSTFLAGS` env var that treats warnings as errors,
but since we currently only run tests on native anyways, I figured it's
alright.