Don't include dev features in release (#391)
`bevy build --release` is not the same as `bevy build --profile release`
currently. If we want Bevy CLI to pick up our release configuration, we
need to explicitly pass `--release`, at which point we might as well
remove the `--profile` argument at the cost of some workflow flexibility
that we aren't using anyways.