Comment out the getrandom fix (#424)
The fix is not strictly needed until a user steps on a mine in the
minefield and their web build breaks. Luckily we check the web build in
CI so they'll notice that there is an issue, and the fix will be in
their `Cargo.toml` ready to be uncommented.